simple_backend/task_tracker/pyproject.toml
[project]
name = "task-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Vladislav Ibraev",email = "vladislavibraev@gmail.com"}
]
readme = "README.md"
requires-python = "==3.12.10"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.11"
pre-commit = "^4.3.0"

[tool.ruff]
line-length = 79
target-version = "py312"
src = ["simple_backend/src/task_tracker"]
extend-exclude = ["migrations", ".venv", "dist", "build"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E203"]
per-file-ignores = { "__init__.py" = ["F401"], "tests/**" = ["S101"] }

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = false
